# fix: use default utc timezone for snowflake queries