<!-- Thanks so much for your PR, your contribution is appreciated! ❤️ -->

Closes: #2826 





### Description:

We alter the snowflake query session to use UTC as the default, this controls the assumed timezone for `TIMESTAMP_TZ` (timestamps with no time zones).

Ideally we should be able to set this parameter as a connection setting but we're blocked by a feature request on the node sdk for snowflake: https://github.com/snowflakedb/snowflake-connector-nodejs/issues/61

For now we prepend every query with an additional query to set the session timezone. I'm sure we take a performance hit but nothing substantial:

**Without** `ALTER SESSION` pre-query
1. `select 1` - 2.128s
2. `select 1` - 0.682s
3. `show parameters like 'timezone'` - 8.588s
4. `show parameters like 'timezone'` - 0.730s

**With** `ALTER SESSION` pre-query
1. `select 1` - 0.908s
2. `select 1` - 1.091s
3. `show parameters like 'timezone'` - 2.301s
4. `show parameters like 'timezone'` - 0.981s